master-agent/handler/requirements.txt
requests
elasticsearch
openai
google-cloud-speech
langgraph
langchain
langchain_openai
langchain-core